feature: migrate to typescript
