--- client/src/pages/sports-grid/components/PlayerInput.js
@@ -1,11 +1,0 @@
-'use client'
-
-import React from "react"
-
-export default function PlayerInput() {
-    return (
-        <div className="player-input">
-            <input></input>
-        </div>
-    )
-}
