+++ client/src/pages/sports-grid/components/PlayerInput.tsx
@@ -1,0 +1,16 @@
+import React from "react"
+
+interface Props {
+    pos: {
+        col: number,
+        row: number,
+    }
+}
+
+export default function PlayerInput({ pos }: Props) {
+    return (
+        <div className="player-input">
+            <input></input>
+        </div>
+    )
+}
