+++ client/src/pages/sports-grid/components/TeamSelector.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import React from 'react';
 import Dropdown from 'react-bootstrap/Dropdown';
 import DropdownButton from 'react-bootstrap/DropdownButton';
