+++ client/src/pages/sports-grid/games/construct-puzzle.ts
@@ -1,8 +1,14 @@
-const seedrandom = require('seedrandom');
+import seedrandom from 'seedrandom';
 
 const PUZZLE_SIZE = 3;
+type Team = {
+    Name: string,
+    Logo: string,
+    Players: string[],
+};
+type TeamsData = Record<string, Team>;
 
-const shuffle = (array, seed=null) => {
+const shuffle = (array: any[], seed=null) => {
     // use today's date as default seed
     const rng = seedrandom(seed || new Date().toJSON().slice(0, 10));
 
@@ -13,13 +19,13 @@ const shuffle = (array, seed=null) => {
     return array;
 }
 
-const hasOverlap = (group1, group2) => {
+const hasOverlap = (group1: Team['Players'], group2: Team['Players']) => {
     const setGroup1 = new Set(group1);
     const intersection = group2.filter(x => setGroup1.has(x));
     return (new Set(intersection)).size > 0;
 }
 
-const constructPuzzle = (teamsJSON) => {
+export const constructPuzzle = (teamsJSON: TeamsData) => {
     const teams = Object.keys(teamsJSON);
     shuffle(teams);
 
@@ -42,5 +48,3 @@ const constructPuzzle = (teamsJSON) => {
         cols
     };
 }
-
-module.exports = { constructPuzzle }
