+++ client/tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "allowJs": true,
+        "skipLibCheck": true,
+        "strict": true,
+        "noEmit": true,
+        "esModuleInterop": true,
+        "module": "esnext",
+        "moduleResolution": "bundler",
+        "resolveJsonModule": true,
+        "outDir": "./build/",
+        "sourceMap": true,
+        "strictNullChecks": true,
+        "jsx": "react"
+
+    },
+    "include": [
+        "./src/"
+    ],
+    "exclude": ["node_modules"]
+}
